+++ modules/__tests__/describeHashSupport.js
@@ -36,6 +36,29 @@ function describeHashSupport(createHistory) {
 
       unlisten = history.listen(execSteps(steps, done))
     })
+
+    it('does not convert PUSH to REPLACE if path does not change', function (done) {
+      const steps = [
+        function (location) {
+          expect(location.pathname).toEqual('/')
+          expect(location.search).toEqual('')
+          expect(location.hash).toEqual('')
+          expect(location.state).toEqual(null)
+          expect(location.action).toEqual(POP)
+
+          history.pushState(null, '/#the-hash')
+        },
+        function (location) {
+          expect(location.pathname).toEqual('/')
+          expect(location.search).toEqual('')
+          expect(location.hash).toEqual('#the-hash')
+          expect(location.state).toEqual(null)
+          expect(location.action).toEqual(PUSH)
+        }
+      ]
+
+      unlisten = history.listen(execSteps(steps, done))
+    })
   })
 }